fast_api_backend/account/pyproject.toml
[project]
name = "fastapi-sample"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "takahiro-maekawa", email = "koma4024@gmail.com" }]
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy>=2.0.35",
    "requests>=2.32.3",
    "async-exit-stack>=1.0.1",
    "async-generator>=1.10",
    "pytest>=8.3.3",
    "flake8>=5.0.4",
    "uvicorn>=0.30.6",
    "sqlalchemy_utils>=0.41.2",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "dependency_injector>=4.42.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["account"]
